Move contacts into main, finish tests
