+++ site/database/tests.py
@@ -28,3 +28,9 @@ class SongTest(TestCase):
         self.assertEqual(song.get_absolute_url(), '/database/test-title/')
         # calling get_absolute_url should set title_slug
         self.assertEqual(song.title_slug, 'test-title')
+
+        song = Song.objects.create(title="Yet Another Test", title_slug="welll")
+        # if has title_slug, nothing changes
+        self.assertEqual(song.title_slug, 'welll')
+        self.assertEqual(song.get_absolute_url(), '/database/welll/')
+        self.assertEqual(song.title_slug, 'welll')
