+++ site/site_settings/urls.py
@@ -3,7 +3,6 @@ from django.contrib import admin
 
 from main.views import *
 from database.views import *
-from contact.views import *
 
 admin.autodiscover()
